module-3/sagemaker-projects-portfolio/requirements.txt
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-asset-awscli-v1==2.2.201 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-asset-kubectl-v20==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-asset-node-proxy-agent-v6==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-aws-lambda-python-alpha==2.122.0a0 ; python_version >= "3.10" and python_version < "4.0"
aws-cdk-lib==2.122.0 ; python_version >= "3.10" and python_version < "4.0"
cattrs==23.2.3 ; python_version >= "3.10" and python_version < "4.0"
cdk-stacksets==0.0.149 ; python_version >= "3.10" and python_version < "4.0"
constructs==10.3.0 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
importlib-resources==6.1.1 ; python_version >= "3.10" and python_version < "4.0"
jsii==1.94.0 ; python_version >= "3.10" and python_version < "4.0"
publication==0.0.3 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "4.0"
